Move BlockListBlockContext out of block.js so that it exists on mobile platforms
